[PP] amount in interest should be typed as BigDecimal

+++ cor/src/main/scala/timetopaytaxpayer/cor/model/Interest.scala
@@ -22,7 +22,7 @@ import play.api.libs.json.{Json, OFormat}
 
 final case class Interest(
     creationDate: Option[LocalDate],
-    amount:       Double
+    amount:       BigDecimal
 )
 
 object Interest {
